Rename wildcard useTheme to useStorybookTheme
