+++ client/wildcard/src/stories/BrandedStory.tsx
@@ -2,9 +2,9 @@ import React from 'react'
 
 import { MemoryRouter, MemoryRouterProps } from 'react-router-dom'
 
-import { WildcardThemeContext } from '../hooks/useWildcardTheme'
+import { WildcardThemeContext } from '../hooks'
 
-import { usePrependStyles, useTheme } from './hooks'
+import { usePrependStyles, useStorybookTheme } from './hooks'
 
 import brandedStyles from '../global-styles/index.scss'
 
@@ -26,7 +26,7 @@ export const BrandedStory: React.FunctionComponent<BrandedProps> = ({
     styles = brandedStyles,
     ...memoryRouterProps
 }) => {
-    const isLightTheme = useTheme()
+    const isLightTheme = useStorybookTheme()
     usePrependStyles('branded-story-styles', styles)
 
     return (
