+++ client/wildcard/src/stories/hooks/index.ts
@@ -1,3 +1,3 @@
 export * from './useChromaticTheme'
 export * from './usePrependStyles'
-export * from './useTheme'
+export * from './useStorybookTheme'
