+++ client/wildcard/src/stories/index.ts
@@ -2,4 +2,4 @@ export * from './BrandedStory'
 export * from './utils/isChromatic'
 export * from './hooks/useChromaticTheme'
 export * from './hooks/usePrependStyles'
-export * from './hooks/useTheme'
+export * from './hooks/useStorybookTheme'
